requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
requests
python-dotenv
pillow          # for image generation
